ai_shopify_dashboard/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
filelock==3.16.0
fsspec==2024.9.0
huggingface-hub==0.24.7
idna==3.8
packaging==24.1
pyactiveresource==2.2.2
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==74.1.2
ShopifyAPI==12.6.0
six==1.16.0
sqlparse==0.5.1
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wheel==0.44.0
